examples/scripts/write_image_as_png_requirements.txt
numpy~=1.26
pillow~=11.2.1